requirements.txt
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
gspread==3.1.0
httplib2==0.12.3
idna==2.8
kiwisolver==1.1.0
numpy==1.16.3
oauth2client==4.1.3
oauthlib==3.0.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
response==0.3.0
rsa==4.0
schedule==0.6.0
slackbot==0.5.3
slacker==0.13.0
uritemplate==3.0.0
urllib3==1.24.2
websocket-client==0.56.0
